Reformat code

+++ matrix_graphics/src/structs.rs
@@ -1,5 +1,3 @@
-use bresenham;
-
 #[derive(Copy, Clone)]
 pub struct Line {
 	pub start_x: f64,
@@ -43,12 +41,12 @@ impl Line {
 			vec![self.rotate_angle.sin(), self.rotate_angle.cos()],
 		];
 		let matrix_start = vec![
-			vec![(self.start_x - self.rotate_center_x) as f64],
-			vec![(self.start_y - self.rotate_center_y) as f64],
+			vec![(self.start_x - self.rotate_center_x)],
+			vec![(self.start_y - self.rotate_center_y)],
 		];
 		let matrix_end = vec![
-			vec![(self.end_x - self.rotate_center_x) as f64],
-			vec![(self.end_y - self.rotate_center_y) as f64],
+			vec![(self.end_x - self.rotate_center_x)],
+			vec![(self.end_y - self.rotate_center_y)],
 		];
 		let res_start = matrix::mult(&matrix_rotate, &matrix_start).unwrap();
 		let res_end = matrix::mult(&matrix_rotate, &matrix_end).unwrap();
